requirements.txt
Django==1.9.2
MySQL-python==1.2.3
django-import-export==0.4.5
tablib==0.11.2

